feat: add setMoment() for using external moment
