+++ package.json
@@ -34,11 +34,6 @@
   ],
   "author": "NHN FE Development Lab <dl_javascript@nhn.com>",
   "license": "MIT",
-  "dependencies": {
-    "eslint-plugin-jest": "^23.6.0",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.27"
-  },
   "devDependencies": {
     "@babel/core": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
@@ -48,8 +43,10 @@
     "eslint-config-prettier": "^6.10.0",
     "eslint-config-tui": "^3.0.0",
     "eslint-loader": "^3.0.3",
+    "eslint-plugin-jest": "^23.6.0",
     "eslint-plugin-prettier": "^3.1.2",
     "jest": "^25.1.0",
+    "moment-timezone": "^0.5.27",
     "prettier": "^1.19.1",
     "safe-umd-webpack-plugin": "^4.0.0",
     "tui-jsdoc-template": "^1.2.2",
