+++ src/index.js
@@ -2,4 +2,5 @@ import LocalDate from './localDate';
 import UTCDate from './utcDate';
 import MomentDate from './momentDate';
 
+export default { LocalDate, UTCDate, MomentDate };
 export { LocalDate, UTCDate, MomentDate };
